+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/aldana/Microverse/calculator/src/index.js":"1","/home/aldana/Microverse/calculator/src/components/App.js":"2","/home/aldana/Microverse/calculator/src/components/Display.js":"3","/home/aldana/Microverse/calculator/src/components/ButtonPanel.js":"4","/home/aldana/Microverse/calculator/src/logic/calculate.js":"5","/home/aldana/Microverse/calculator/src/components/Button.js":"6","/home/aldana/Microverse/calculator/src/logic/operate.js":"7"},{"size":237,"mtime":1611042746415,"results":"8","hashOfConfig":"9"},{"size":656,"mtime":1611042661771,"results":"10","hashOfConfig":"9"},{"size":444,"mtime":1611042661771,"results":"11","hashOfConfig":"9"},{"size":1369,"mtime":1611042661771,"results":"12","hashOfConfig":"9"},{"size":1070,"mtime":1610697005699,"results":"13","hashOfConfig":"9"},{"size":433,"mtime":1611042661771,"results":"14","hashOfConfig":"9"},{"size":410,"mtime":1610696371835,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"xab6bz",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},"/home/aldana/Microverse/calculator/src/index.js",[],"/home/aldana/Microverse/calculator/src/components/App.js",[],"/home/aldana/Microverse/calculator/src/components/Display.js",[],"/home/aldana/Microverse/calculator/src/components/ButtonPanel.js",[],"/home/aldana/Microverse/calculator/src/logic/calculate.js",[],["31","32","33","34","35","36","37"],"/home/aldana/Microverse/calculator/src/components/Button.js",[],"/home/aldana/Microverse/calculator/src/logic/operate.js",[],{"ruleId":"38","replacedBy":"39"},{"ruleId":"40","replacedBy":"41"},{"ruleId":"42","replacedBy":"43"},{"ruleId":"44","replacedBy":"45"},{"ruleId":"46","replacedBy":"47"},{"ruleId":"48","replacedBy":"49"},{"ruleId":"50","replacedBy":"51"},"jsx-a11y/accessible-emoji",[],"lines-around-directive",["52"],"no-spaced-func",["53"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements","func-call-spacing"]
+[{"/home/aldana/Microverse/calculator/src/index.js":"1","/home/aldana/Microverse/calculator/src/components/App.js":"2","/home/aldana/Microverse/calculator/src/logic/calculate.js":"3","/home/aldana/Microverse/calculator/src/components/shared/Navbar.js":"4","/home/aldana/Microverse/calculator/src/logic/operate.js":"5","/home/aldana/Microverse/calculator/src/components/ButtonPanel.js":"6","/home/aldana/Microverse/calculator/src/components/Display.js":"7","/home/aldana/Microverse/calculator/src/components/Button.js":"8","/home/aldana/Microverse/calculator/src/components/Calculator.js":"9"},{"size":277,"mtime":1611046475770,"results":"10","hashOfConfig":"11"},{"size":230,"mtime":1611047112956,"results":"12","hashOfConfig":"11"},{"size":1070,"mtime":1610697005699,"results":"13","hashOfConfig":"11"},{"size":285,"mtime":1611046377700,"results":"14","hashOfConfig":"11"},{"size":410,"mtime":1610696371835,"results":"15","hashOfConfig":"11"},{"size":1369,"mtime":1611045810540,"results":"16","hashOfConfig":"11"},{"size":444,"mtime":1611045810540,"results":"17","hashOfConfig":"11"},{"size":433,"mtime":1611045810540,"results":"18","hashOfConfig":"11"},{"size":748,"mtime":1611046676054,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"xab6bz",{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/aldana/Microverse/calculator/src/index.js",[],"/home/aldana/Microverse/calculator/src/components/App.js",[],"/home/aldana/Microverse/calculator/src/logic/calculate.js",[],["39","40","41","42","43","44","45"],"/home/aldana/Microverse/calculator/src/components/shared/Navbar.js",[],"/home/aldana/Microverse/calculator/src/logic/operate.js",[],"/home/aldana/Microverse/calculator/src/components/ButtonPanel.js",[],"/home/aldana/Microverse/calculator/src/components/Display.js",[],"/home/aldana/Microverse/calculator/src/components/Button.js",[],"/home/aldana/Microverse/calculator/src/components/Calculator.js",[],{"ruleId":"46","replacedBy":"47"},{"ruleId":"48","replacedBy":"49"},{"ruleId":"50","replacedBy":"51"},{"ruleId":"52","replacedBy":"53"},{"ruleId":"54","replacedBy":"55"},{"ruleId":"56","replacedBy":"57"},{"ruleId":"58","replacedBy":"59"},"jsx-a11y/accessible-emoji",[],"lines-around-directive",["60"],"no-spaced-func",["61"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements","func-call-spacing"]
